The ROOT directory is not what you think it is
The ROOT directory used to be the subdirectory that contains this script - currently common/scripts. The envs and docs subdirectories are created in the actual root directory of this repository, two levels above common/scripts. This happened because of this part of the code:
	ROOT=$(get_script_dir)
	cd "$ROOT/../.."
Modify get_script_dir() to return the actual root directory and correctly set ROOT.